% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ItemGaps}
\alias{ItemGaps}
\title{ItemGaps: Workhorse of moste mixed type questions.}
\usage{
ItemGaps(Titel, ID, Question, Answer, Points, fqti, fqpl, Author)
}
\arguments{
\item{title}{Name des Fragenpools}

\item{ID}{laufender Index, welcher die Schnittstelle zwischen qpl- und qti-Datei regelt}

\item{Question}{bestimmt den eigentlichen Text einer Aufgabenstellung, z.B. "Berechnen sie das arith. Mittel"}

\item{Answer}{bestimmt die Antwortmöglichkeiten -> z.B. bei num. Aufg. das Ergebnisintervall der ind. Aufgabe -> arith. Mittel zu spez. Tabelleninhalten
ist eine Liste mit Listenelementen den entsprechenden Lücken. variablen in Liste siehe Details}

\item{Points}{reglet die Punktevergabe der Aufgaben eines Fragenpools -> idR identische Punkteanzahl für dieselbe Aufgabenstellung}

\item{Author}{Hier kann der jeweilige Name des Erstellers eines Fragenpools mitgegeben werden}

\item{fqti}{File-Handler}

\item{fqpl}{File-Handler}
}
\description{
ItemGaps: Workhorse of moste mixed type questions.
}
\details{
\code{type} = num,drop,text,
\code{min,max,true} für grenzen beim Typ num und den wahren Wert
\code{labels,true} für Typ text und drop, Vektor mit labels als Antwortmöglichkeiten, true Booleen welche Antwort ist richtig.
\code{shuffle} "Yes" oder "No": Ob die Antwortmöglichkeiten zufällig angeordnet werden sollen.
}

